+++ src/python/cloud/gdrive_constants.py
@@ -12,7 +12,7 @@ import os
 # ---------------------------------------------------------------------------
 # Version
 # ---------------------------------------------------------------------------
-VERSION = "1.18.0"
+VERSION = "1.18.1"
 
 # ---------------------------------------------------------------------------
 # API / OAuth
